+++ DemoApp/SwiftCacheDemoApp.swift
@@ -1,0 +1,18 @@
+//
+//  SwiftCacheDemoApp.swift
+//  SwiftCache Demo
+//
+//  Created by Sudhir Gadhvi
+//
+
+import SwiftUI
+
+@main
+struct SwiftCacheDemoApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}
+